Refactor tqec.circuit.schedule into submodules
