+++ src/tqec/circuit/schedule/circuit_test.py
@@ -1,20 +1,11 @@
-import random
-
 import pytest
 import stim
 
 from tqec.circuit.moment import Moment
 from tqec.circuit.qubit import GridQubit
 from tqec.circuit.qubit_map import QubitMap
-from tqec.circuit.schedule import (
-    Schedule,
-    ScheduledCircuit,
-    ScheduleException,
-    merge_scheduled_circuits,
-    relabel_circuits_qubit_indices,
-    remove_duplicate_instructions,
-)
-from tqec.exceptions import TQECException, TQECWarning
+from tqec.circuit.schedule import Schedule, ScheduledCircuit, ScheduleException
+from tqec.exceptions import TQECException
 
 _VALID_SCHEDULED_CIRCUITS = [
     stim.Circuit(""),
@@ -28,88 +19,6 @@ _VALID_SCHEDULED_CIRCUITS = [
         "TICK\nCX 0 1\nTICK\nM 0 1"
     ),
 ]
-
-
-def test_schedule_construction() -> None:
-    Schedule([])
-    Schedule([0])
-    Schedule([0, 2, 4, 6])
-    Schedule(list(range(100000)))
-
-    with pytest.raises(ScheduleException):
-        Schedule([-2, 0, 1])
-    with pytest.raises(ScheduleException):
-        Schedule([0, 1, 1, 2])
-    with pytest.raises(ScheduleException):
-        Schedule([1, 0])
-
-
-def test_schedule_from_offset() -> None:
-    Schedule.from_offsets([])
-    Schedule.from_offsets([0])
-    Schedule.from_offsets([0, 2, 4, 6])
-    Schedule.from_offsets(list(range(100000)))
-
-    with pytest.raises(ScheduleException):
-        Schedule.from_offsets([-2, 0, 1])
-    with pytest.raises(ScheduleException):
-        Schedule.from_offsets([0, 1, 1, 2])
-    with pytest.raises(ScheduleException):
-        Schedule.from_offsets([1, 0])
-
-
-def test_schedule_len() -> None:
-    assert len(Schedule([])) == 0
-    assert len(Schedule([0])) == 1
-    assert len(Schedule([0, 2, 4, 6])) == 4
-    assert len(Schedule(list(range(100000)))) == 100000
-
-
-def test_schedule_getitem() -> None:
-    with pytest.raises(IndexError):
-        Schedule([])[0]
-    with pytest.raises(IndexError):
-        Schedule([])[-1]
-    assert Schedule([0])[0] == 0
-    assert Schedule([0, 2, 4, 6])[2] == 4
-    assert Schedule([0, 2, 4, 6])[-1] == 6
-    sched = Schedule(list(range(100000)))
-    for i in (random.randint(0, 99999) for _ in range(100)):
-        assert sched[i] == i
-
-
-def test_schedule_iter() -> None:
-    assert list(Schedule([])) == []
-    assert list(Schedule([0, 5, 6])) == [0, 5, 6]
-    assert list(Schedule(list(range(100000)))) == list(range(100000))
-
-
-def test_schedule_append_schedule() -> None:
-    sched = Schedule(list(range(10)))
-    sched.append_schedule(sched)
-    assert sched.schedule == list(range(20))
-
-
-def test_schedule_insert() -> None:
-    schedule = Schedule([0, 3])
-    with pytest.raises(ScheduleException):
-        schedule.insert(0, -1)
-    assert schedule.schedule == [0, 3]
-    schedule.insert(1, 1)
-    assert schedule.schedule == [0, 1, 3]
-    schedule.insert(-1, 2)
-    assert schedule.schedule == [0, 1, 2, 3]
-
-
-def test_schedule_append() -> None:
-    schedule = Schedule([0, 3])
-    with pytest.raises(ScheduleException):
-        schedule.append(-1)
-    assert schedule.schedule == [0, 3]
-    schedule.append(4)
-    assert schedule.schedule == [0, 3, 4]
-    schedule.append(5)
-    assert schedule.schedule == [0, 3, 4, 5]
 
 
 def test_scheduled_circuit_construction() -> None:
@@ -342,100 +251,3 @@ def test_scheduled_circuit_filter_by_qubit() -> None:
     filtered_circuit = circuit.filter_by_qubits([GridQubit(0, 0), GridQubit(0, 1)])
     assert filtered_circuit.get_circuit() == stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")
     assert filtered_circuit.schedule == Schedule([0])
-
-
-def test_remove_duplicate_instructions() -> None:
-    instructions: list[stim.CircuitInstruction] = list(
-        iter(stim.Circuit("H 0 0 0 2 0 0 0 0 1 1 2 2 0 0 0 3"))
-    )  # type: ignore
-    expected_instructions = set(iter(stim.Circuit("H 0 1 2 3")))  # type: ignore
-    assert (
-        set(remove_duplicate_instructions(instructions, frozenset(["H"])))
-        == expected_instructions
-    )
-    with pytest.warns(TQECWarning):
-        # Several H gates are overlapping, which means that the returned instruction
-        # list is not a valid Moment, which should raise a warning.
-        assert remove_duplicate_instructions(instructions, frozenset()) == instructions
-
-    instructions = list(iter(stim.Circuit("H 0 1 2 3 0 1 2 3\nM 4 5 6 4 5 6 4 5 6")))  # type: ignore
-    with pytest.warns(TQECWarning):
-        assert set(
-            remove_duplicate_instructions(instructions, frozenset(["M"]))
-        ) == set(iter(stim.Circuit("H 0 1 2 3 0 1 2 3\nM 4 5 6")))
-    with pytest.warns(TQECWarning):
-        assert set(
-            remove_duplicate_instructions(instructions, frozenset(["H"]))
-        ) == set(iter(stim.Circuit("H 0 1 2 3\nM 4 5 6 4 5 6 4 5 6")))
-    assert set(
-        remove_duplicate_instructions(instructions, frozenset(["H", "M"]))
-    ) == set(iter(stim.Circuit("H 0 1 2 3\nM 4 5 6")))
-
-
-def test_relabel_circuits_qubit_indices() -> None:
-    # Any qubit target not defined by a QUBIT_COORDS instruction should raise
-    # an exception.
-    with pytest.raises(KeyError):
-        relabel_circuits_qubit_indices(
-            [
-                ScheduledCircuit.from_circuit(stim.Circuit("H 0 1 2")),
-                ScheduledCircuit.from_circuit(stim.Circuit("H 0 1 2")),
-            ]
-        )
-    with pytest.raises(KeyError):
-        relabel_circuits_qubit_indices(
-            [
-                ScheduledCircuit.from_circuit(
-                    stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")
-                ),
-                ScheduledCircuit.from_circuit(stim.Circuit("H 0")),
-            ]
-        )
-    circuits, q2i = relabel_circuits_qubit_indices(
-        [
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")),
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(1, 1) 0\nX 0")),
-        ]
-    )
-    assert q2i == QubitMap({0: GridQubit(0, 0), 1: GridQubit(1, 1)})
-    assert len(circuits) == 2
-    assert circuits[0].get_circuit() == stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")
-    assert circuits[1].get_circuit() == stim.Circuit("QUBIT_COORDS(1, 1) 1\nX 1")
-
-
-def test_merge_scheduled_circuits() -> None:
-    # Any qubit target not defined by a QUBIT_COORDS instruction should raise
-    # an exception.
-    _circuits, _qubit_map = relabel_circuits_qubit_indices(
-        [
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")),
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(1, 1) 0\nX 0")),
-        ]
-    )
-    circuit = merge_scheduled_circuits(_circuits, _qubit_map)
-    assert circuit.get_circuit() == stim.Circuit(
-        "QUBIT_COORDS(0, 0) 0\nQUBIT_COORDS(1, 1) 1\nH 0\nX 1"
-    )
-
-    circuit = merge_scheduled_circuits(
-        [
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")),
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")),
-        ],
-        global_qubit_map=QubitMap({0: GridQubit(0, 0)}),
-        mergeable_instructions=["H"],
-    )
-    assert circuit.get_circuit() == stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0")
-
-    _circuits, _qubit_map = relabel_circuits_qubit_indices(
-        [
-            ScheduledCircuit.from_circuit(
-                stim.Circuit("QUBIT_COORDS(0, 0) 0\nH 0\nTICK\nM 0"), [0, 2]
-            ),
-            ScheduledCircuit.from_circuit(stim.Circuit("QUBIT_COORDS(1, 1) 0\nX 0"), 1),
-        ]
-    )
-    circuit = merge_scheduled_circuits(_circuits, _qubit_map)
-    assert circuit.get_circuit() == stim.Circuit(
-        "QUBIT_COORDS(0, 0) 0\nQUBIT_COORDS(1, 1) 1\nH 0\nTICK\nX 1\nTICK\nM 0"
-    )
