+++ src/tqec/circuit/schedule/exception.py
@@ -1,0 +1,5 @@
+from tqec.exceptions import TQECException
+
+
+class ScheduleException(TQECException):
+    pass
